+++ src/MiscWebApp/Sample.cs
@@ -1,0 +1,5 @@
+namespace MiscWebApp;
+
+public class Sample {
+	public int Value { get; set; }
+}
